+++ 2023/day_10/task_1/Cargo.toml
@@ -1,8 +1,18 @@
 [package]
-name = "task_1"
+name = "solution_2023_10_1"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "solution_2023_10_1"
+path = "src/lib.rs"
+crate-type = ["rlib", "cdylib"]
+
+[profile.release]
+opt-level = "z"
+lto = true
+codegen-units = 1
+panic = "abort"
 
 [dependencies]
+wasm-bindgen = "0.2.100"
