+++ 2023/day_12/task_2/Cargo.toml
@@ -1,9 +1,19 @@
 [package]
-name = "task_2"
+name = "solution_2023_12_2"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "solution_2023_12_2"
+path = "src/lib.rs"
+crate-type = ["rlib", "cdylib"]
+
+[profile.release]
+opt-level = "z"
+lto = true
+codegen-units = 1
+panic = "abort"
 
 [dependencies]
-memoize = "0.4.1"
+wasm-bindgen = "0.2.100"
+memoize = "0.5.1"
